refactor: Settings 및 진단 UI를 feature 구조로 분리
