+++ frontend/src/features/settings/index.ts
@@ -1,0 +1,1 @@
+export { SettingsDrawer } from "./SettingsDrawer";
